requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
fastapi==0.73.0
h11==0.13.0
httptools==0.3.0
hurry.filesize==0.9
idna==3.3
Pillow==9.0.0
protobuf==3.19.4
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
